feat: fixed some bugs and allowed users to query from the pipe icon
