+++ src/flows/components/header/Submit.tsx
@@ -1,12 +1,5 @@
 // Libraries
-import React, {
-  FC,
-  MouseEvent,
-  useMemo,
-  useContext,
-  useState,
-  useEffect,
-} from 'react'
+import React, {FC, MouseEvent, useContext, useEffect, useMemo} from 'react'
 import {
   Dropdown,
   IconFont,
@@ -16,29 +9,16 @@ import {
   List,
   ButtonGroup,
 } from '@influxdata/clockface'
-import {useDispatch} from 'react-redux'
 
 import {SubmitQueryButton} from 'src/timeMachine/components/SubmitQueryButton'
 import {QueryContext} from 'src/flows/context/query'
-import {FlowContext} from 'src/flows/context/flow.current'
 import {RunModeContext, RunMode} from 'src/flows/context/runMode'
-import {ResultsContext} from 'src/flows/context/results'
 import {notify} from 'src/shared/actions/notifications'
-import {PIPE_DEFINITIONS} from 'src/flows'
+import {FlowContext} from 'src/flows/context/flow.current'
 
 // Utils
-import {event} from 'src/cloud/utils/reporting'
 import {cancelAllRunningQueries} from 'src/timeMachine/actions/queries'
-
-// Constants
-import {
-  notebookRunSuccess,
-  notebookRunFail,
-} from 'src/shared/copy/notifications'
-import {PROJECT_NAME} from 'src/flows'
-
-// Types
-import {RemoteDataState} from 'src/types'
+import {PIPE_DEFINITIONS} from 'src/flows'
 
 // Styles
 import 'src/flows/components/header/Submit.scss'
@@ -46,100 +26,34 @@ import 'src/flows/components/header/Submit.scss'
 const fakeNotify = notify
 
 export const Submit: FC = () => {
-  const dispatch = useDispatch()
-  const {query, generateMap} = useContext(QueryContext)
-  const {runMode, setRunMode} = useContext(RunModeContext)
   const {flow} = useContext(FlowContext)
-  const {add, update} = useContext(ResultsContext)
-  const [isLoading, setLoading] = useState(RemoteDataState.NotStarted)
-
-  const hasQueries = useMemo(() => {
-    return !!flow.data.all
-      .map(p => PIPE_DEFINITIONS[p.type])
-      .filter(p => p && (p.family === 'transform' || p.family === 'inputs'))
-      .length
-  }, [flow.data])
-
-  // Returns a string to avoid array deep comparison woes
-  // Triggers a re-run any time any aggregate function is changed
-  const aggregateOfAggregates = useMemo(() => {
-    return flow.data.all
-      .map(p => p.aggregateFunction?.name)
-      .filter(p => !!p)
-      .join('')
-  }, [flow.data])
+  const {runMode, setRunMode} = useContext(RunModeContext)
+  const {isLoading, queryAll} = useContext(QueryContext)
 
   const _range = useMemo(
     () => `${flow.range.lower} to ${flow.range.upper || 'now'}`,
     [flow.range]
   )
 
+  const hasQueries = useMemo(() => {
+    return !!flow.data.all
+      .filter(p => {
+        // this condition ensures that empty metric selectors are filtered out
+        if (p?.type === 'metricSelector') {
+          return !!p?.bucket
+        }
+        return true
+      })
+      .map(p => PIPE_DEFINITIONS[p.type])
+      .filter(p => p && (p.family === 'transform' || p.family === 'inputs'))
+      .length
+  }, [flow.data])
+
   useEffect(() => {
     if (hasQueries) {
-      submit()
+      queryAll()
     }
-  }, [_range, hasQueries, aggregateOfAggregates]) // eslint-disable-line react-hooks/exhaustive-deps
-
-  const forceUpdate = (id, data) => {
-    try {
-      update(id, data)
-    } catch (_e) {
-      add(id, data)
-    }
-  }
-
-  const submit = () => {
-    if (isLoading === RemoteDataState.Loading) {
-      return
-    }
-
-    const map = generateMap(runMode === RunMode.Run)
-
-    if (!map.length) {
-      return
-    }
-
-    event('Flow Submit Button Clicked')
-    setLoading(RemoteDataState.Loading)
-
-    flow.data.allIDs.forEach(pipeID => {
-      flow.meta.update(pipeID, {loading: RemoteDataState.Loading})
-    })
-
-    Promise.all(
-      map.map(stage => {
-        return query(stage.text)
-          .then(response => {
-            stage.instances.forEach(pipeID => {
-              flow.meta.update(pipeID, {loading: RemoteDataState.Done})
-              forceUpdate(pipeID, response)
-            })
-          })
-          .catch(e => {
-            stage.instances.forEach(pipeID => {
-              forceUpdate(pipeID, {
-                error: e.message,
-              })
-              flow.meta.update(pipeID, {loading: RemoteDataState.Error})
-            })
-          })
-      })
-    )
-      .then(() => {
-        event('run_notebook_success', {runMode})
-        dispatch(notify(notebookRunSuccess(runMode, PROJECT_NAME)))
-
-        setLoading(RemoteDataState.Done)
-      })
-      .catch(e => {
-        event('run_notebook_fail', {runMode})
-        dispatch(notify(notebookRunFail(runMode, PROJECT_NAME)))
-
-        // NOTE: this shouldn't fire, but lets wrap it for completeness
-        setLoading(RemoteDataState.Error)
-        throw e
-      })
-  }
+  }, [_range, hasQueries]) // eslint-disable-line react-hooks/exhaustive-deps
 
   const DropdownButton = (
     active: boolean,
@@ -151,9 +65,9 @@ export const Submit: FC = () => {
           text={runMode}
           className="flows--submit-button"
           icon={IconFont.Play}
-          submitButtonDisabled={!hasQueries}
+          submitButtonDisabled={hasQueries === false}
           queryStatus={isLoading}
-          onSubmit={submit}
+          onSubmit={queryAll}
           onNotify={fakeNotify}
           queryID=""
           cancelAllRunningQueries={cancelAllRunningQueries}
