+++ src/flows/context/query.tsx
@@ -1,6 +1,5 @@
-import React, {FC, useContext, useMemo, useEffect} from 'react'
+import React, {FC, useContext, useMemo, useEffect, useState} from 'react'
 import {useDispatch, useSelector} from 'react-redux'
-import {AppState, ResourceType, RemoteDataState} from 'src/types'
 import {parse} from 'src/external/parser'
 import {runQuery} from 'src/shared/apis/query'
 import {getWindowVars} from 'src/variables/utils/getWindowVars'
@@ -11,6 +10,8 @@ import {getBuckets} from 'src/buckets/actions/thunks'
 import {getSortedBuckets} from 'src/buckets/selectors'
 import {getStatus} from 'src/resources/selectors'
 import {FlowContext} from 'src/flows/context/flow.current'
+import {RunModeContext, RunMode} from 'src/flows/context/runMode'
+import {ResultsContext} from 'src/flows/context/results'
 import {fromFlux} from '@influxdata/giraffe'
 import {event} from 'src/cloud/utils/reporting'
 import {FluxResult} from 'src/types/flows'
@@ -19,20 +20,34 @@ import {
   generateHashedQueryID,
   setQueryByHashID,
 } from 'src/timeMachine/actions/queries'
+import {notify} from 'src/shared/actions/notifications'
 
+// Constants
+import {
+  notebookRunSuccess,
+  notebookRunFail,
+} from 'src/shared/copy/notifications'
+import {PROJECT_NAME} from 'src/flows'
+
+// Types
+import {AppState, ResourceType, RemoteDataState} from 'src/types'
 interface Stage {
   text: string
   instances: string[]
 }
 
 export interface QueryContextType {
-  query: (text: string) => Promise<FluxResult>
   generateMap: (withSideEffects?: boolean) => Stage[]
+  isLoading: RemoteDataState
+  query: (text: string) => Promise<FluxResult>
+  queryAll: () => void
 }
 
 export const DEFAULT_CONTEXT: QueryContextType = {
-  query: () => Promise.resolve({} as FluxResult),
   generateMap: () => [],
+  isLoading: RemoteDataState.NotStarted,
+  query: (_: string) => Promise.resolve({} as FluxResult),
+  queryAll: () => {},
 }
 
 export const QueryContext = React.createContext<QueryContextType>(
@@ -81,11 +96,14 @@ const findOrgID = (text, buckets) => {
 
 export const QueryProvider: FC = ({children}) => {
   const {flow} = useContext(FlowContext)
+  const {runMode} = useContext(RunModeContext)
+  const {add, update} = useContext(ResultsContext)
   const variables = useSelector((state: AppState) => getVariables(state))
   const buckets = useSelector((state: AppState) => getSortedBuckets(state))
   const bucketsLoadingState = useSelector((state: AppState) =>
     getStatus(state, ResourceType.Buckets)
   )
+  const [isLoading, setLoading] = useState(RemoteDataState.NotStarted)
 
   const dispatch = useDispatch()
 
@@ -189,6 +207,71 @@ export const QueryProvider: FC = ({children}) => {
           error: null,
         }
       })
+      .catch(error => {
+        console.error('error: ', error)
+      })
+  }
+
+  const forceUpdate = (id, data) => {
+    try {
+      update(id, data)
+    } catch (_e) {
+      add(id, data)
+    }
+  }
+
+  const queryAll = () => {
+    if (isLoading === RemoteDataState.Loading) {
+      return
+    }
+
+    const map = generateMap(runMode === RunMode.Run)
+
+    if (!map.length) {
+      return
+    }
+
+    console.log('map: ', map)
+
+    event('Flow Submit Button Clicked')
+    setLoading(RemoteDataState.Loading)
+
+    Promise.all(
+      map.map(stage => {
+        stage.instances.forEach(pipeID => {
+          flow.meta.update(pipeID, {loading: RemoteDataState.Loading})
+        })
+        return query(stage.text)
+          .then(response => {
+            stage.instances.forEach(pipeID => {
+              flow.meta.update(pipeID, {loading: RemoteDataState.Done})
+              forceUpdate(pipeID, response)
+            })
+          })
+          .catch(e => {
+            stage.instances.forEach(pipeID => {
+              forceUpdate(pipeID, {
+                error: e.message,
+              })
+              flow.meta.update(pipeID, {loading: RemoteDataState.Error})
+            })
+          })
+      })
+    )
+      .then(() => {
+        event('run_notebook_success', {runMode})
+        dispatch(notify(notebookRunSuccess(runMode, PROJECT_NAME)))
+
+        setLoading(RemoteDataState.Done)
+      })
+      .catch(e => {
+        event('run_notebook_fail', {runMode})
+        dispatch(notify(notebookRunFail(runMode, PROJECT_NAME)))
+
+        // NOTE: this shouldn't fire, but lets wrap it for completeness
+        setLoading(RemoteDataState.Error)
+        // throw e
+      })
   }
 
   if (!flow?.range || bucketsLoadingState !== RemoteDataState.Done) {
@@ -196,7 +279,7 @@ export const QueryProvider: FC = ({children}) => {
   }
 
   return (
-    <QueryContext.Provider value={{query, generateMap}}>
+    <QueryContext.Provider value={{query, generateMap, isLoading, queryAll}}>
       {children}
     </QueryContext.Provider>
   )
