+++ src/flows/pipes/MetricSelector/AggregateWindowSelector.tsx
@@ -7,6 +7,7 @@ import {Dropdown, IconFont, Icon} from '@influxdata/clockface'
 // Contexts
 import {PipeContext} from 'src/flows/context/pipe'
 import {FlowContext} from 'src/flows/context/flow.current'
+import {QueryContext} from 'src/flows/context/query'
 
 // Constants
 import {FUNCTIONS, QueryFn} from 'src/timeMachine/constants/queryBuilder'
@@ -17,6 +18,8 @@ import {millisecondsToDuration} from 'src/shared/utils/duration'
 const AggregateFunctionSelector: FC = () => {
   const {flow} = useContext(FlowContext)
   const {data, update} = useContext(PipeContext)
+  const {queryAll} = useContext(QueryContext)
+
   const selectedFunction = data?.aggregateFunction || FUNCTIONS[0]
 
   const windowPeriod = flow?.range?.windowPeriod
@@ -31,8 +34,9 @@ const AggregateFunctionSelector: FC = () => {
         function: aggregateFunction.name,
       })
       update({aggregateFunction})
+      queryAll()
     },
-    [update]
+    [update, queryAll]
   )
 
   const menuItems = (
