+++ src/flows/shared/ResizerHeader.tsx
@@ -5,9 +5,15 @@ import classnames from 'classnames'
 // Components
 import {Icon, IconFont} from '@influxdata/clockface'
 
+// Context
+import {PipeContext} from 'src/flows/context/pipe'
+import {QueryContext} from 'src/flows/context/query'
+
+// Utils
+import {event} from 'src/cloud/utils/reporting'
+
 // Types
 import {Visibility} from 'src/types/flows'
-import {PipeContext} from 'src/flows/context/pipe'
 
 interface Props {
   visibility: Visibility
@@ -34,6 +40,7 @@ const ResizerHeader: FC<Props> = ({
   toggleVisibilityEnabled,
 }) => {
   const {readOnly} = useContext(PipeContext)
+  const {queryAll} = useContext(QueryContext)
   const glyph = visibility === 'visible' ? IconFont.EyeOpen : IconFont.EyeClosed
   const className = classnames('panel-resizer--header', {
     'panel-resizer--header__multiple-controls':
@@ -41,9 +48,14 @@ const ResizerHeader: FC<Props> = ({
     [`panel-resizer--header__${visibility}`]: resizingEnabled && visibility,
   })
 
+  const handleSubmit = () => {
+    event('Query All button clicked from inside Pipe')
+    queryAll()
+  }
+
   if (!resizingEnabled) {
     return (
-      <div className={className}>
+      <div className={className} onClick={handleSubmit}>
         <Icon glyph={emptyIcon} className="panel-resizer--vis-toggle" />
       </div>
     )
